@@ -1,39 +1,44 @@
 {
   "name": "cogniflect",
-  "version": "0.1.1",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "concurrently": "^4.1.0",
-    "immutable": "^3.8.2",
+    "concurrently": "^5.0.0",
+    "immutable": "^4.0.0-rc.12",
     "merge": ">=1.2.1",
-    "nodemon": "^1.18.7",
-    "react": "^16.6.3",
-    "react-bootstrap": "^0.32.4",
-    "react-dom": "^16.6.3",
-    "react-redux": "^5.1.1",
-    "react-router": "^4.3.1",
-    "react-router-bootstrap": "^0.24.4",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.1",
-    "redux": "^4.0.1",
+    "nodemon": "^1.19.4",
+    "react": "^16.10.2",
+    "react-bootstrap": "^1.0.0-beta.5",
+    "react-dom": "^16.10.2",
+    "react-redux": "^7.1.1",
+    "react-router": "^5.1.2",
+    "react-router-bootstrap": "^0.25.0",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "^3.2.0",
+    "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
-    "redux-saga": "^0.16.2",
-    "socket.io": "^2.2.0",
-    "socket.io-client": "^2.2.0"
+    "redux-saga": "^1.1.1",
+    "socket.io": "^2.3.0",
+    "socket.io-client": "^2.3.0",
+    "winston": "^3.2.1"
   },
   "devDependencies": {
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "husky": "1.2.0",
-    "react-test-renderer": "^16.6.3"
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
+    "husky": "3.0.9",
+    "react-test-renderer": "^16.10.2"
   },
   "scripts": {
-    "build": "react-scripts build",
+    "build": "concurrently \"npm:build-*\"",
+    "build-socketserver": "babel src/lib -d dist -s --ignore \"src/**/*.test.js\"",
+    "build-webapp": "react-scripts build",
     "check-updates": "ncu",
-    "coverage": "react-scripts test --env=jsdom --coverage",
+    "coverage": "react-scripts test --env=jsdom --coverage --watchAll=false",
     "eject": "react-scripts eject",
     "start": "concurrently \"npm:start-*\"",
-    "start-socketserver": "nodemon -w src/server.js src/server.js",
+    "start-socketserver": "nodemon -w dist/server.js dist/index.js",
     "start-webapp": "react-scripts start",
     "test": "react-scripts test --env=jsdom"
   },
