@@ -1,15 +1,13 @@
 import React from 'react';
 import {Button, Col, Row} from 'react-bootstrap';
 import {LinkContainer} from 'react-router-bootstrap';
-import { connect } from 'react-redux';
-import actions from '../actions';
 
 const StartOver = ({children, questions, startOverHandler}) => (
   <Row>
     <Col xs={12}>
         <p>
           <LinkContainer to="/">
-            <Button onClick={startOverHandler(questions)} bsStyle="warning">Start Over</Button>
+            <Button onClick={startOverHandler(questions)} variant="warning">Start Over</Button>
           </LinkContainer>
           {children}
         </p>
@@ -17,21 +15,4 @@ const StartOver = ({children, questions, startOverHandler}) => (
   </Row>
 );
 
-const mapStateToProps = (state) => ({
-  questions: state.appState.get('questions')
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  startOverHandler: (questions) => (e) => {
-    dispatch(actions.set(['assessmentType'], 'individual'));
-    dispatch(actions.set(['teamCode'], ''));
-    dispatch(actions.set(['userTeamCode'], ''));
-    dispatch(actions.set(['questionCursor'], 0));
-    dispatch(actions.set(['questions'], questions.map((question) => {
-      return question.set('myAnswer', null).set('ourAnswer', null);
-    })));
-  }
-});
-
-export const DumbComponent = StartOver;
-export const SmartComponent = connect(mapStateToProps, mapDispatchToProps)(StartOver);
+export default StartOver;
