@@ -1,9 +1,6 @@
 import React from 'react';
-import {SmartComponent as StartOver, DumbComponent} from './StartOver';
-import { createStore } from 'redux';
-import reducers from '../reducers';
+import StartOver from './StartOver';
 import {fromJS} from 'immutable';
-import ShallowRenderer from 'react-test-renderer/shallow';
 
 const initialState = {
   appState: fromJS({
@@ -20,45 +17,9 @@ const initialState = {
   })
 };
 
-const store = createStore(reducers, initialState);
-  
-describe('smart component', () => {
-  var renderedComp;
-  beforeEach(() => {
-    store.dispatch = jest.fn();
-    const renderer = new ShallowRenderer();
-    renderedComp = renderer.render(
-      <StartOver store={store}/>
-    );
-  });
-
+describe('component/StartOver', () => {
   it('matches snapshot', () => {
-    expect(renderedComp).toMatchSnapshot()
-  });
-
-  it('dispatches "startOverHandler"', () => {
-    renderedComp.props.startOverHandler(initialState.appState.get('questions'))();
-    expect(store.dispatch.mock.calls.length).toBe(5);
-  });
-
-  afterEach(() => {
-    store.dispatch.mockRestore();
-  });
-});
-
-describe('dumb component', () => {
-  var renderedComp,
-      startOverHandler = jest.fn(() => () => true);
-  beforeEach(() => {
-    const renderer = new ShallowRenderer();
-    renderedComp = renderer.render(
-      <DumbComponent {...initialState} startOverHandler={startOverHandler}>
-        <p>Hello, World!</p>
-      </DumbComponent>
-    );
-  });
-  
-  it('without crashing', () => {
-    expect(renderedComp).toMatchSnapshot();
+    const component = StartOver({...initialState, startOverHandler:jest.fn()})
+    expect(component).toMatchSnapshot();
   });
 });
