@@ -1,9 +1,8 @@
 import React from 'react';
-import {SmartComponent as Synchronize, DumbComponent} from './Synchronize';
+import Synchronize from './Synchronize';
 import { createStore } from 'redux';
 import reducers from '../reducers';
 import {fromJS} from 'immutable';
-import ShallowRenderer from 'react-test-renderer/shallow';
 
 const initialState = {
   appState: fromJS({
@@ -15,41 +14,10 @@ const initialState = {
   })
 };
 
-const store = createStore(reducers, initialState);
-
-describe('smart component', () => {
-  var renderedComp;
-  beforeEach(() => {
-    const renderer = new ShallowRenderer();
-    renderedComp = renderer.render(
-      <Synchronize store={store}/>
-    );
-  });
-
+describe('components/Synchronize', () => {
   it('matches snapshot', () => {
-    expect(renderedComp).toMatchSnapshot()
-  });
-  
-  it('dispatches "continueHandler"', (done) => {
-    store.subscribe((action) => {
-      expect(store.getState().appState.get('assessmentType')).toEqual('team');
-      done();
-    });
-    renderedComp.props.continueHandler();
-  });
-});
-
-describe('dumb component', () => {
-  var renderer;
-  beforeEach(() => {
-    renderer = new ShallowRenderer();
-  });
-  
-  it('matches snapshot', () => {
-    const renderedComp = renderer.render(
-      <DumbComponent {...initialState}/>
-    );
-    expect(renderedComp).toMatchSnapshot();
+    const component = Synchronize(initialState)
+    expect(component).toMatchSnapshot();
   });
   
   it('enables "Continue"', () => {
@@ -57,9 +25,7 @@ describe('dumb component', () => {
       appState: initialState.appState
         .setIn(['team', 'synced'], 2)
     };
-    const renderedComp = renderer.render(
-      <DumbComponent {...enabledState} proceed={true}/>
-    );
-    expect(renderedComp).toMatchSnapshot();
+    const component = Synchronize({...enabledState, proceed:true })
+    expect(component).toMatchSnapshot();
   });
 });
